requirements.txt
openai
python-telegram-bot
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
